requirements.txt
boto3==1.34.144
botocore==1.34.144
tkinterdnd2==0.3.0
python-dotenv==1.0.0